pyproject.toml
[project]
name = "mindspace"
version = "1.5.1"
description = "Hierarchical Knowledge Agent"
readme = "QUICKSTART.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py",
    "google-genai",
    "GitPython",
    "pyyaml",
    "beautifulsoup4",
    "httpx",
    "pageindex",
    "openviking"
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["mindspace*"]
